promoted ExportDialog to activity
