+++ res/layout/export_stats.xml
@@ -11,11 +11,7 @@
 		android:gravity="center"
 		android:layout_width="fill_parent"
 		android:orientation="vertical"
-		android:layout_marginTop="40dp"
-		android:layout_marginBottom="40dp"
-		android:layout_marginLeft="10dp"
-		android:layout_marginRight="10dp"
-		android:background="@drawable/modal_dialog">
+		android:background="@drawable/app_background">
 
 		<ListView
 			android:id="@+id/list_view_id"
@@ -26,9 +22,8 @@
 			android:layout_centerInParent="true"
 			android:divider="#fff"
 			android:cacheColorHint="#fff"
-			android:layout_height="wrap_content"
+			android:layout_height="fill_parent"
 			android:layout_above="@+id/export_dialog_buttons"
-			android:layout_below="@+id/export_dialog_title"
 			android:layout_margin="2dp" />
 
 
@@ -92,7 +87,7 @@
 					    android:layout_height="wrap_content"
 					    android:layout_gravity="center"
 					    android:padding="5dp"
-					    android:text="@string/close"
+					    android:text="@string/cancel"
 					    android:textColor="#fff"
 					    android:textStyle="bold" />
