+++ res/values/strings.xml
@@ -120,6 +120,7 @@
 	<string name="export_no_app">No app selected, can\'t export!</string>
 	<string name="export_started">Export started</string>
     <string name="export_finished">Export finished</string>
+    <string name="export_failed_to_load_stats">Failed to load app stats.</string>
     <string name="export_saved_to">Saved to</string>
     <string name="share">Share</string>
