+++ src/com/github/andlyticsproject/util/Utils.java
@@ -58,6 +58,10 @@ public final class Utils {
 		}
 	}
 
+	public static <P, T extends AsyncTask<P, ?, ?>> void execute(T task) {
+		execute(task, (P[]) null);
+	}
+
 	@SuppressLint("NewApi")
 	public static <P, T extends AsyncTask<P, ?, ?>> void execute(T task, P... params) {
 		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
